fix: show local store upgrade alert directly
