+++ Sources/Application/AppController.swift
@@ -97,10 +97,6 @@ import PostHog
 
     func applicationWillFinishLaunching(_ notification: Notification) {
         bindChromiumBridgeIfNeeded()
-        NotificationCenter.default.addObserver(self,
-                                               selector: #selector(localStoreRequiresNewerApp(_:)),
-                                               name: .localStoreRequiresNewerApp,
-                                               object: nil)
 
         // Register defaults before any settings are read.
         UserDefaultsRegistration.registerDefaults()
@@ -270,23 +266,6 @@ import PostHog
     @objc private func loginCompleted(_ notification: Notification) {
         Task { @MainActor in
             self.flushPendingOpenURLsAwaitingLoginStatus()
-        }
-    }
-
-    @objc private func localStoreRequiresNewerApp(_ notification: Notification) {
-        Task { @MainActor in
-            let alert = NSAlert()
-            alert.messageText = NSLocalizedString(
-                "Update Phi to Open Local Data",
-                comment: "Local store compatibility alert - title when the local database was opened by a newer app version"
-            )
-            alert.informativeText = NSLocalizedString(
-                "This version of Phi cannot open local browser data that was updated by a newer version. Install the latest Phi version and try again.",
-                comment: "Local store compatibility alert - body when a newer app is required to read local data"
-            )
-            alert.alertStyle = .warning
-            alert.addButton(withTitle: NSLocalizedString("OK", comment: "Generic - OK button to dismiss an alert"))
-            alert.runModal()
         }
     }
